pyproject.toml
[tool.poetry]
name = "py-etl-con-duckdb"
version = "0.1.0"
description = ""
authors = ["crisolmedo07 <crisolmedo07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gdown = "^5.2.0"
duckdb = "^1.0.0" 
streamlit = "^1.37.1"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
